+++ app/src/main/java/com/dvsnier/wrapper/TransferStationWrapper.java
@@ -11,22 +11,22 @@ import com.dvsnier.base.flavor.crash.TestCrashHandleActivity;
 import com.dvsnier.base.flavor.sql.TestSQLActivity;
 import com.dvsnier.bean.ComponentBean;
 import com.dvsnier.constant.IAdapterType;
+import com.dvsnier.test.common.permission.TestPermissionActivity;
 import com.dvsnier.test.view.affinity.AffinityActivity;
-import com.dvsnier.test.widget.theme.TestThemeActivity;
-import com.dvsnier.testAIDL.TestAIDLActivity;
 import com.dvsnier.test.widget.animator.TestAnimator;
 import com.dvsnier.test.widget.coordinator.TestCoordinatorLayout2Activity;
 import com.dvsnier.test.widget.coordinator.TestCoordinatorLayout3Activity;
 import com.dvsnier.test.widget.coordinator.TestCoordinatorLayoutActivity;
+import com.dvsnier.test.widget.recycleview.TestRecyclerActivity;
+import com.dvsnier.test.widget.scroll.TestScrollActivity;
+import com.dvsnier.test.widget.surfaceview.TestSurfaceActivity;
+import com.dvsnier.test.widget.theme.TestThemeActivity;
+import com.dvsnier.testAIDL.TestAIDLActivity;
 import com.dvsnier.testEventBus.TestEventBusActivity;
 import com.dvsnier.testGreenDao.TestGreenDaoActivity;
 import com.dvsnier.testImage.TestImageActivity;
 import com.dvsnier.testOkhttp.TestOkhttpActivity;
-import com.dvsnier.testPermission.TestPermissionActivity;
-import com.dvsnier.test.widget.recycleview.TestRecyclerActivity;
-import com.dvsnier.test.widget.scroll.TestScrollActivity;
 import com.dvsnier.testSpeechRecognition.TestSpeechRecognitionActivity;
-import com.dvsnier.test.widget.surfaceview.TestSurfaceActivity;
 import com.dvsnier.testTouchEvent.TestTouchEventActivity;
 import com.dvsnier.testXUtils.TestXUtilsActivity;
 import com.orhanobut.logger.Logger;
