+++ research/common/src/main/res/layout/layout_common_button.xml
