+++ research/common/src/main/res/values/bools.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <bool name="debug">true</bool>
+    <bool name="debug_permission">true</bool>
 </resources>
