+++ research/interceptor/build.gradle
@@ -28,6 +28,7 @@ dependencies {
 
     api project(':research:view')
     api project(':research:widget')
+    api project(':research:common')
 
     testImplementation "junit:junit:$rootProject.junit_version"
